backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
paynow==1.0.6
python-dotenv==1.0.0
